moved accounts from home component to new Accounts component, added Accounts component to home component
