+++ src/components/account-details/AccountDetails.tsx
@@ -2,10 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useContext } from 'react';
 import { UserContext } from '../../context/user.context';
 import { Transaction } from '../../models/Transaction';
-import {
-  apiGetAccount,
-  apiGetTransactions,
-} from '../../remote/banking-api/account.api';
+import { apiGetAccounts, apiGetTransactions } from '../../remote/banking-api/account.api';
 import Navbar from '../navbar/Navbar';
 import { useNavigate } from 'react-router-dom';
 import './AccountDetails.css';
@@ -19,12 +16,13 @@ import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
 import Button from '@mui/material/Button';
+import { useAppSelector, useAppDispatch } from '../../app/hooks';
 
 export default function AccountDetails() {
   const navigate = useNavigate();
   const { user } = useContext(UserContext);
   const [transaction, setTransactions] = useState<Transaction[]>([]);
-  const [account, setAccount] = React.useState<Account>();
+  const currentAccount = useAppSelector((state) => state.account.currentAccount);
   let txnForm = <></>;
 
   useEffect(() => {
@@ -33,28 +31,28 @@ export default function AccountDetails() {
     }
     const fetchData = async () => {
       if (user) {
-        const resultAcct = await apiGetAccount(user.id);
-        setAccount(resultAcct.payload);
-        const result = await apiGetTransactions(resultAcct.payload.id);
+        // const resultAcct = await apiGetAccounts(user.id);
+        // setAccount(resultAcct.payload);
+        const result = await apiGetTransactions(currentAccount?.id);
         setTransactions(result.payload.reverse());
       }
     };
     fetchData();
   }, [user, navigate]);
 
-  useEffect(() => {
-    (async () => {
-      if (user) {
-        const resultAcct = await apiGetAccount(user.id);
-        setAccount(resultAcct.payload);
-      }
-    })();
-  }, [transaction, user]);
+  // useEffect(() => {
+  //   (async () => {
+  //     if (user) {
+  //       const resultAcct = await apiGetAccounts(user.id);
+  //       setAccount(resultAcct.payload);
+  //     }
+  //   })();
+  // }, [transaction, user]);
 
-  if (account) {
+  if (currentAccount) {
     txnForm = (
       <CreateTransactionForm
-        accountId={account!.id}
+        accountId={currentAccount?.id}
         afterUpsert={(result) => setTransactions([result, ...transaction])}
       />
     );
@@ -65,9 +63,9 @@ export default function AccountDetails() {
       <Navbar />
       <div className="account-wrap">
         <div className="account-details">
-          <h2>{account?.name}</h2>
-          <h3>{account?.description}</h3>
-          <h1>${account?.balance}</h1>
+          <h2>{currentAccount.name}</h2>
+          <h3>{currentAccount.description}</h3>
+          <h1>{currentAccount.balance}</h1>
           <Button
             onClick={() => {
               navigate('/');
